chore: resolve circular deps
